feat: update deployment manifest to use GCR for audio-backend image
This commit updates the deployment manifest to use a new Docker image location for the audio-backend application. The new image location is now stored in Google Container Registry (GCR) instead of Artifact Registry.

This change was made to improve the reliability and performance of the audio-backend application by leveraging the benefits of GCR, such as faster image pull speeds and enhanced security features.